Disable PostBuildEvent for StartUpDummy project.
Unfortunately, the PostBuildEvent failed in many situations, such as:
* User was using the Microsoft Store version of Paint.NET
* User did not install Paint.NET into the solution's root directory
* User had Paint.NET open (or other file-in-use errors causing copy to fail)

This commit also updates the path to reference the corresponding files.
This should allow building binaries, even when using Microsoft Store
version of Paint.NET v4.1.1 (or even if not installed on the build machine).
This is a step towards automated commit checks (e.g., TravisCI builds).