CS0659 - Matrix overrides Object.Equals(object o), but does not override Object.GetHashCode()
This is an imperfect solution to the problem, caused by the Matrix
being a mutable reference type.  It's generally not a great idea
to override Equals() on reference types, and strongly advised against
for mutable reference types.

Unfortunately, this is not code I wrote, so will need a deeper
analysis to determine if appropriate to ever change....